+++ src/components/EmptyStateCard/assets/assets.defaults.ts
@@ -1,0 +1,7 @@
+import type { EmptyStateCardAssetDefaultProps } from "./assets.types";
+
+export const emptyStateCardAssetDefaultProps: EmptyStateCardAssetDefaultProps =
+  {
+    icon: "Square",
+    color: "primary",
+  };