+++ src/components/EmptyStateCard/assets/index.ts
@@ -1,0 +1,2 @@
+export * from "./EmptyBox.asset";
+export * from "./TaskSleepingCat.asset";