+++ src/components/EmptyStateCard/index.ts
@@ -1,0 +1,2 @@
+export * from "./EmptyStateCard.component";
+export type * from "./EmptyStateCard.types";